+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 ### Bug Fixes
   1. [#1065](https://github.com/influxdata/chronograf/pull/1065): Save and Cancel edits to a Dashboard name
+  2. [#1069](https://github.com/influxdata/chronograf/pull/1069): Graphs are no longer editable from a Host Page
 
 ## v1.2.0-beta6 [2017-03-24]